OneDrive/Desktop/ITEMS/DJANGO/ASHIQ-PORTFOLIO/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.3
botocore==1.27.3
cffi==1.15.1
cryptography==38.0.1
Django==4.0.5
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-rest-knox==4.2.0
django-storages==1.12.3
djangorestframework==3.14.0
gunicorn==20.1.0
jmespath==1.0.0
Pillow==9.1.1
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.5
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
